micros/language-detection/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
lingua-language-detector==2.1.1
